+++ lib/smart_core/initializer/plugins/thy_types.rb
@@ -14,9 +14,17 @@ class SmartCore::Initializer::Plugins::ThyTypes < SmartCore::Initializer::Plugin
         '::Thy does not exist or "thy" gme is not loaded'
       ) unless const_defined?('::Thy')
 
+      # NOTE: require necessary dependencies
       require 'date'
+
+      # NOTE: add thy-types type system implementation
       require_relative 'thy_types/errors'
       require_relative 'thy_types/thy_types'
+
+      # NOTE: register thy-types type system
+      SmartCore::Initializer::TypeSystem.register(
+        :thy_types, SmartCore::Initializer::TypeSystem::ThyTypes
+      )
     end
   end
 end
