+++ spec/spec_helper.rb
@@ -3,7 +3,7 @@
 require 'simplecov'
 
 SimpleCov.formatter = SimpleCov::Formatter::HTMLFormatter
-SimpleCov.minimum_coverage(100)
+SimpleCov.minimum_coverage(100) if !!ENV['FULL_TEST_COVERAGE_CHECK']
 SimpleCov.start do
   enable_coverage :branch
   add_filter 'spec'
